requirements.txt
streamlit==1.39.0
tensorflow==2.20.0
numpy==1.26.4
pillow==10.4.0
protobuf==4.25.3
